% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_phiprime.R
\name{add_phiprime}
\alias{add_phiprime}
\title{Add Phi' isolines to a Linf/K plot}
\usage{
add_phiprime(gridsize = 20, ...)
}
\arguments{
\item{gridsize}{bla}

\item{...}{bla}
}
\value{
contour layer
}
\description{
Add Phi' isolines to a Linf/K plot
}
