% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv.FLIBM.R
\name{adv.FLIBM}
\alias{adv.FLIBM}
\title{Advance FLIBM stock}
\usage{
adv.FLIBM(obj, years = NULL, seasons = NULL, monitor = TRUE)
}
\arguments{
\item{obj}{object of class FLIBM}

\item{years}{character vector. years to cycle for advancement.
Defaults to full length of dimension. Vector should follow chronological order.}

\item{seasons}{character vector. seasons to cycle for advancement.
Defaults to full length of dimension. Vector should follow chronological order.
If shorter than a full year, then the year argument should be for a single year.}

\item{monitor}{logical. Should progression be printed.}
}
\value{
FLIBM object
}
\description{
Advance FLIBM stock
}
\examples{

stk <- create.FLIBM(years=ac(2000:2009))
stk$rec$params$season_wt[] <- c(0,0,0.25,1,0.25,0,0,0,0,0,0,0)
dimnames(stk$stock.l@stock.n)['year']
nrow(stk$inds)
stk$stock.a@stock.n[,'2000',,,,]

stk <- adv.FLIBM(obj = stk, years = ac(2000:2009))
plot(stk$stock.a@stock.n)

# plot length-frequency
DIM <- dim(stk$stock.l@catch.n)
DIMNAMES <- dimnames(stk$stock.l@catch.n)

lfq <- FLIBM2lfq(stk)

pal <- colorRampPalette(c("grey30",5,7,2), bias=1.5)
with(lfq, image(x=dates, y=midLengths, z=t(catch), col=pal(100)))


# numbers by age at start of the year
stk$stock.a@range[c("minfbar", "maxfbar")] <- c(1,4)
stk.yr <- simplifySeason(stk)
stk.yr <- stk.yr[2:dim(stk.yr)[1],]
plot(stk.yr)

# Biomass
plot(stk$stock.a@stock.n[,,,1,,]*stk$stock.a@stock.wt[,,,1,,])

# catch curve
tmp <- stk.yr@catch.n[,"2009"]
tmp
plot(1:6, log(c(tmp)))
df <- data.frame(logN = log(c(tmp))[1:4], age = 1:4)
fit <- lm(logN ~ age, df)
fit
abline(fit, col=2)


}
