% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_t.R
\name{calc_t}
\alias{calc_t}
\title{calc_t}
\usage{
calc_t(par = NULL, Ltnew = NULL, Lt = NULL, t = NULL,
  tincr = 0.01)
}
\arguments{
\item{par}{bla}

\item{Ltnew}{bla}

\item{Lt}{bla}

\item{t}{bla}

\item{tincr}{bla}
}
\value{
numeric year decimal value
}
\description{
calc_t
}
\examples{

plot(1)


}
