% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFM.R
\name{calcFM}
\alias{calcFM}
\title{Calculate FM based on target catch}
\usage{
calcFM(
  obj,
  Ctarget = 1000,
  FMStart = NULL,
  lower = 0.01,
  upper = 2,
  year = NULL,
  seed = 1,
  abstol = 5
)
}
\arguments{
\item{obj}{bla}

\item{Ctarget}{bla}

\item{FMStart}{bla}

\item{lower}{bla}

\item{upper}{bla}

\item{year}{bla}

\item{seed}{bla}

\item{abstol}{numeric. Absolute tolerance for quota}

\item{unit}{bla}

\item{season}{bla}

\item{area}{bla}

\item{iter}{bla}
}
\value{
value
}
\description{
Calculate FM based on target catch
}
\examples{
\donttest{
set.seed(1)
stk <- create.FLIBM(year=ac(2000:2009))
range(stk$stock.a)[c("minfbar", "maxfbar")] <- c(1,3)

stk <- spinup.FLIBM(stk, nyearsmax = 20, nyearsslope = 5)
stk <- adv.FLIBM(obj = stk, year = ac(2000:2004))
plot(stk$stock.a@stock.n)
plot(simplifySeason(stk))


# optim version
system.time(res <- calcFM(obj = stk, Ctarget = 1200, year = "2005", seed = 1))
res$par # best val
res$value # difference to target

# check
set.seed(1)
stk2 <- stk
stk2$harvest$params$FM <- c(res$par)
stk2 <- adv.FLIBM(obj = stk2, year = "2005")

# total catches
apply(stk2$stock.a@catch.n * stk2$stock.a@catch.wt, 2, sum, na.rm=TRUE)

# fishing mortality
apply(harvest(stk2$stock.a),1:2, sum, na.rm=TRUE)
fbar(simplifySeason(stk2))



### OTHER OPTION WITH GA  - not cxurrently used ###

system.time(res <- calcFM(obj = stk, Ctarget = 1200, year = "2005",
  seed = 1, popSize = 10, maxiter = 20, abstol = 5))

summary(res)


# check
set.seed(1)
stk2 <- stk
stk2$harvest$params$FM <- c(summary(res)$sol)
stk2 <- adv.FLIBM(obj = stk2, year = "2005")

# total catches
apply(stk2$stock.a@catch.n * stk2$stock.a@catch.wt, 2, sum, na.rm=TRUE)

# fishing mortality
apply(harvest(stk2$stock.a),1:2, sum, na.rm=TRUE)
fbar(simplifySeason(stk2))
}

}
