% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.FLIBM.R
\name{cohort.FLIBM}
\alias{cohort.FLIBM}
\title{Simulate a single cohort}
\usage{
cohort.FLIBM(obj = NULL, ssbfec = 1e+06, FM = 0.3,
  years = dimnames(obj$stock.a)$year, seed = 1, monitor = FALSE,
  return.FLIBM = FALSE)
}
\arguments{
\item{obj}{FLIBM object.}

\item{ssbfec}{numeric value. Single value describing constant spawning stock
biomass (or fecundity), used in determining recruitment with \code{obj$rec}
during the first year of simulation.}

\item{FM}{numeric value. Fishing mortality to be applied.}

\item{years}{character vector. Years to use in simulation
(from \code{ dimnames(obj$stock.a@stock.n)} )}

\item{seed}{seed value for random number reproducibility (Default: 1).
When seed = NULL, a random seed is used during each FM run.}
}
\value{
a list
}
\description{
Simulate a single cohort
}
\examples{


data(stkMed)

# obj$rec parameters should result in high recruitment for the given
# ssbfec setting (e.g. many thousands of individuals will result in
# smoother ypr pattern)
stkMed$rec$params$rmax <- 1e5
obj <- stkMed

res <- cohort.FLIBM(obj, ssbfec = 5e5, FM = 0.3, seed = 1,
  return.FLIBM = TRUE)

plot(res$obj$stock.a@stock.n[,ac(1980:1995)])

stkYr <- simplifySeason(res$obj)
stkYr <- stkYr[-1,ac(1980:2000)]
plot(stkYr)


}
