% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.FLIBM.R
\name{create.FLIBM}
\alias{create.FLIBM}
\title{Create initial FLIBM object}
\usage{
create.FLIBM(lengths = seq(0, 85), ages = 0:6, years = ac(2000:2009),
  units = "unique", seasons = ac(1:12), areas = "unique",
  iters = "1", n.units = "1e3", wt.units = "kg", name = "synStock",
  desc = "A synthetic stock", inds = NULL)
}
\arguments{
\item{lengths}{vector. Lengths to define the length extent of
the length-based FLStock object (lower bin extent).}

\item{ages}{vector. Ages to define the age extent of
the age-based FLStock object (lower bin extent).}

\item{years}{vector. Years to define the 'year' extent of
the FLStock object.}

\item{units}{vector. Unitss to define the 'unit' extent of
the FLStock object.}

\item{seasons}{vector. Seasons to define the 'season' extent of
the FLStock object.}

\item{areas}{vector. Areas to define the 'area' extent of
the FLStock object.}

\item{iters}{vector. Iters to define the 'iter' extent of
the FLStock object.}

\item{n.units}{character. Units decribing the scale for numbers
used in the FLStock object. If value is higher than "1" (e.g. "1e3"),
then individuals in the IBM are also considered as
"super-individuals", representing a aggregate of individuals
with like attributes. (Default="1e3")}

\item{wt.units}{character. Units decribing the scale for individual
weights used in the FLStock object, and in the IBM. (Default="kg").}

\item{name}{character. Name of stock.}

\item{desc}{character. Description of stock.}

\item{inds}{data.frame. (Super-) individuals (rows) of the current
population state with their attributes (column variables).}
}
\value{
FLIBM object
}
\description{
Create initial FLIBM object
}
\examples{
stk <- create.FLIBM()
summary(stk)

}
