% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmsy.FLIBM.R
\name{fmsy.FLIBM}
\alias{fmsy.FLIBM}
\title{Estimate maximum sustainable yeild (Fmsy) of an FLIBM object}
\usage{
fmsy.FLIBM(obj = NULL, FMs = seq(0, 1, 0.1),
  years = dimnames(obj$stock.a)$year,
  yearsCompare = dimnames(obj$stock.a)$year, parallel = TRUE,
  no_cores = detectCores() - 1, clusterType = "PSOCK",
  outfile = "output.txt", seed = 1, resDir = NULL, cleanup = FALSE)
}
\arguments{
\item{obj}{bla}

\item{FMs}{bla}

\item{years}{bla}

\item{yearsCompare}{bla}

\item{parallel}{bla}

\item{no_cores}{bla}

\item{clusterType}{bla}

\item{outfile}{bla}

\item{seed}{bla}

\item{resDir}{bla}

\item{cleanup}{bla}
}
\value{
data.frame
}
\description{
bla
}
\examples{

data(stkMed)
stkMed$rec$params$rmax <- 1e5
resdf <- fmsy.FLIBM(stkMed, years = ac(1980:2009),
  yearsCompare = ac(2000:2009), no_cores = 3,
  FMs = seq(0,0.5,length.out = 12),
  resDir = file.path(getwd(), "tmp"),
  parallel = TRUE
)
refptPlot(resdf, ypr = FALSE, spar=0.15, ps=12)

# ypr plot should fail because 'Rrecr' variable is missing
# refptPlot(resdf, ypr = TRUE, spar=0.4, ps=12)




}
