% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inds.R
\name{grow.inds}
\alias{grow.inds}
\title{Grow individuals}
\usage{
grow.inds(obj, year = NA, season = NA)
}
\arguments{
\item{obj}{bla}

\item{year}{bla}

\item{season}{bla}
}
\value{
FLIBM object
}
\description{
Grow individuals
}
