% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{LQ}
\alias{LQ}
\title{Calculate length quantile for a given logistic ogive}
\usage{
LQ(L50, wqs, targQ = 0.95)
}
\arguments{
\item{L50}{Length at prob 50\%}

\item{wqs}{width between 25\% and 75\% quantiles}

\item{targQ}{target probability quantile (Default = 95\%)}
}
\description{
Calculate length quantile for a given logistic ogive
}
