% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inds.R
\name{remove.inds}
\alias{remove.inds}
\title{Remove dead individuals}
\usage{
remove.inds(obj, year = NA, season = NA, purgeProb = 0.05)
}
\arguments{
\item{obj}{bla}

\item{year}{bla}

\item{season}{bla}

\item{purgeProb}{probability of purging empty inds during time step
(between 0 and 1)}
}
\value{
FLIBM object
}
\description{
Remove dead individuals
}
