% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inds.R
\name{reproduce.inds}
\alias{reproduce.inds}
\title{Reproduce individuals}
\usage{
reproduce.inds(obj, year = NA, season = NA)
}
\arguments{
\item{obj}{bla}

\item{year}{bla}

\item{season}{bla}
}
\value{
FLIBM object
}
\description{
Reproduce individuals
}
