% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_hash.R
\name{seed_hash}
\alias{seed_hash}
\title{print seed}
\usage{
seed_hash(prefix = "Hash of seed: ")
}
\arguments{
\item{prefix}{of choice}
}
\value{
text
}
\description{
print seed
}
\examples{
cat(seed_hash())

}
