% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLIBM.R, R/simplifySeason.R
\name{simplifySeason}
\alias{simplifySeason}
\title{Collapse season in age-based FLStock in FLIBM object}
\usage{
simplifySeason(obj)

simplifySeason(obj)
}
\arguments{
\item{obj}{}

\item{obj}{}
}
\value{
FLStock object with season dimension collapsed

FLStock object with season dimension collapsed
}
\description{
The \code{simplifySeason} function takes an FLIBM object, which
typically includes a season dimension, and returns an age-based FLStock
object with the season dimension collapsed. The FLCore function
\code{simplify} is first called in order to set up the yearly object,
followed by additional changes to slots. Numbers slots are based on values
from the beginning of the year, while average weight slots are weighted
averages by year with numbers at season as the weighting. \code{m.spwn}
and \code{harvest.spwn} are based on the seasonal spawning pattern weights
as defined by \code{obj$rec$params$season_wt}.

The \code{simplifySeason} function takes an FLIBM object, which
typically includes a season dimension, and returns an age-based FLStock
object with the season dimension collapsed. The FLCore function
\code{simplify} is first called in order to set up the yearly object,
followed by additional changes to slots. Numbers slots are based on values
from the beginning of the year, while average weight slots are weighted
averages by year with numbers at season as the weighting. \code{m.spwn}
and \code{harvest.spwn} are based on the seasonal spawning pattern weights
as defined by \code{obj$rec$params$season_wt}.
}
\examples{
data(stk1)
plot(stk1$stock.a@stock.n)

stkYr <- simplifySeason(stk1)
stkYr <- stkYr[ac(1:range(stkYr)["max"]),] # remove age 0
plot(stkYr)


data(stk1)
plot(stk1$stock.a@stock.n)

stkYr <- simplifySeason(stk1)
stkYr <- stkYr[ac(1:range(stkYr)["max"]),] # remove age 0
plot(stkYr)


}
