% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup.FLIBM.R
\name{spinup.FLIBM}
\alias{spinup.FLIBM}
\title{Spin up of inds in FLIBM object}
\usage{
spinup.FLIBM(obj, nyearsmax = 20, nyearsslope = 5, unit = NULL,
  area = NULL, iter = NULL, monitor = TRUE)
}
\arguments{
\item{obj}{FLIBM object}

\item{nyearsmax}{numeric. Maximum number of years for spin-up (Default=20)}

\item{nyearsslope}{numeric. Number of years for used to determin if spawning-
stock biomass is in equilibrium (via linear model of log SSB) (Default=5).}

\item{unit}{character vector. units to cycle through for spin-up.
Defaults to full length of dimension.}

\item{area}{character vector. areas to cycle through for spin-up.
Defaults to full length of dimension.}

\item{iter}{character vector. iters to cycle through for spin-up.
Defaults to full length of dimension.}

\item{monitor}{logical. Should progression be printed.}
}
\value{
FLIBM object
}
\description{
Spin up of inds in FLIBM object
}
\examples{
stk <- create.FLIBM()
round(sum(stk$inds[[1]][[1]][[1]]$wt * stk$inds[[1]][[1]][[1]]$mat,
  na.rm = TRUE),3)
nrow(stk$inds[[1]][[1]][[1]])




}
