% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inds.R
\name{update.inds}
\alias{update.inds}
\title{Update variables of individuals}
\usage{
update.inds(obj, year = NA, unit = NA, season = NA, area = NA,
  iter = NA)
}
\arguments{
\item{obj}{bla}

\item{year}{bla}

\item{unit}{bla}

\item{season}{bla}

\item{area}{bla}

\item{iter}{bla}
}
\value{
FLIBM object
}
\description{
Update variables of individuals
}
