% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/war.FLIBM.R
\name{war.FLIBM}
\alias{war.FLIBM}
\title{Calculate width at recruitment (WAR) of a single cohort}
\usage{
war.FLIBM(obj = NULL, ssbfec = 1e+06, FM = 0.2,
  years = dimnames(obj$stock.a)$year, qs = c(0.05, 0.95), minN = 100,
  monitor = FALSE, plot = TRUE)
}
\arguments{
\item{obj}{FLIBM object}

\item{ssbfec}{numeric value. Single value describing constant spawning stock
biomass (or fecundity), used in determining recruitment with \code{obj$rec}
during the first year of simulation.}

\item{FM}{numeric value ofmaximum fishing mortality to be applied
(Default: `FM = 0.2`).}

\item{years}{character vector. Years to use in simulation
(from \code{ dimnames(obj$stock.a@stock.n)} )}

\item{qs}{numeric vector of length. Definesthe lower and upper quantiles to
use in calculating `war` (Default: `qs = c(0.05, 0.95)`)}

\item{minN}{numeric value. Defines the minimum number of individuals
required to estimate quantiles at a given time.}

\item{monitor}{logical. Should progression be printed.}

\item{plot}{logical. Should summary plot be drawn.}
}
\value{
list. Contains a summary data.frame with statistics over time
  over time (`$df`), and estimates of width at recruitment (`$war`),
  moving average (`MA`), and length at first capture (`L50`).
}
\description{
The function `war.FLIBM` simulates a single cohort
  (i.e. spawning year) and estimates the width of the cohort length
  distribution through time (width between defined quantiles).
  The width at recruitment (`$war`) is then the width of the cohort when
  the lower quantile surpasses the length of recruitment to the fishery
  (`L50`). Results provide a way of estimating the `MA` setting that should
  be used within the function \code{\link[TropFishR]{lfqRestructure}}
  of the TropFishR package.
}
\examples{

## load data
data(stkMed)
stkMed$rec$params$rmax <- 1e4

## war analysis
set.seed(1111)
res <- war.FLIBM(obj = stkMed, qs = c(0.1, 0.9),
  FM = 0.2, years = ac(1980:1985),
  monitor = TRUE, plot = TRUE)

## estimated values
res$war # width at recruitment
res$MA # moving average setting given current bin size
res$L50 # length at first capture

}
