% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypr.FLIBM.R
\name{ypr.FLIBM}
\alias{ypr.FLIBM}
\title{Perform a yeild-per-recruit (YPR) analysis on an FLIBM object}
\usage{
ypr.FLIBM(obj = NULL, ssbfec = 1e+06, FMs = seq(0, 1, 0.05),
  years = dimnames(obj$stock.a)$year, parallel = TRUE,
  no_cores = detectCores() - 1, clusterType = "PSOCK",
  outfile = "output.txt", seed = 1)
}
\arguments{
\item{obj}{FLIBM object.}

\item{ssbfec}{numeric value. Single value describing constant spawning stock
biomass (or fecundity), used in determining recruitment with \code{obj$rec}
during the first year of simulation.}

\item{FMs}{numeric vector. Sequence of maximum fishing mortalities to
be applied.}

\item{years}{character vector. Years to use in simulation
(from \code{ dimnames(obj$stock.a@stock.n)} )}

\item{parallel}{logical. Simulations run in parallel.}

\item{no_cores}{numeric value. Number of cores to use when
\code{parallel = TRUE} (Default: \code{no_cores = detectCores() - 1})}

\item{clusterType}{(Default: `clusterType = "PSOCK"`)}

\item{outfile}{character. Text file name (Default: `outfile = "output.txt"`)
which will records the progress of the permutation completions.}

\item{seed}{seed value for random number reproducibility (Default: 1).
When seed = NULL, a random seed is used during each FM run.}
}
\value{
data.frame
}
\description{
bla
}
\examples{

data(stkMed)

# obj$rec parameters should result in high recruitment for the given
# ssbfec setting (e.g. many thousands of individuals will result in
# smoother ypr pattern)
stkMed$rec$params$rmax <- 1e4



system.time(
  resdf <- ypr.FLIBM(obj = stkMed, ssbfec = 1e6, parallel = TRUE,
    FMs = seq(0,0.6,len=12), seed = 1)
)

plot(SSB ~ FM, resdf, t = "o")
plot(Catch/Recr ~ FM, resdf, t = "o")

# with refptPlot
calcRefpts(resdf, ypr=TRUE, spar=0.3)


}
