% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inds.R
\name{die.inds}
\alias{die.inds}
\title{Die individuals}
\usage{
die.inds(obj, year = NA, season = NA)
}
\arguments{
\item{obj}{bla}

\item{year}{bla}

\item{season}{bla}
}
\value{
FLIBM object
}
\description{
Die individuals
}
