% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmsy.FLIBM.R
\name{fmsy.FLIBM}
\alias{fmsy.FLIBM}
\title{Estimate maximum sustainable yeild (Fmsy) of an FLIBM object}
\usage{
fmsy.FLIBM(obj = NULL, FMs = seq(0, 1, 0.1),
  years = dimnames(obj$stock.a)$year,
  yearsCompare = dimnames(obj$stock.a)$year, parallel = TRUE,
  no_cores = detectCores() - 1, clusterType = "PSOCK",
  outfile = "output.txt", seed = 1, resDir = tempdir(),
  cleanup = TRUE)
}
\arguments{
\item{obj}{bla}

\item{FMs}{bla}

\item{years}{bla}

\item{yearsCompare}{bla}

\item{parallel}{bla}

\item{no_cores}{bla}

\item{clusterType}{bla}

\item{outfile}{bla}

\item{seed}{bla}

\item{resDir}{path to intermediate objects (default = \code{tempdir()}).
Change directory if objects should be kept. Otherwise temporary directory
is used, which is automatically deleted following the R session.}

\item{cleanup}{logical. Default TRUE.}
}
\value{
data.frame
}
\description{
bla
}
\examples{

data(stkMed)
stkMed$rec$params$rmax <- 1e2
resdf <- fmsy.FLIBM(stkMed, years = ac(2000:2009),
  yearsCompare = ac(2005:2009), no_cores = 3,
  FMs = seq(0,1,length.out = 6),
  cleanup = TRUE,
  parallel = TRUE
)
calcRefpts(resdf, ypr = FALSE, spar=0.15)



}
