% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inds.R
\name{grow.inds}
\alias{grow.inds}
\title{Grow individuals}
\usage{
grow.inds(obj, year = NA, unit = NA, season = NA, area = NA,
  iter = NA)
}
\arguments{
\item{obj}{bla}

\item{year}{bla}

\item{unit}{bla}

\item{season}{bla}

\item{area}{bla}

\item{iter}{bla}
}
\value{
FLIBM object
}
\description{
Grow individuals
}
