% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inds.R
\name{record.inds}
\alias{record.inds}
\title{Record state of population}
\usage{
record.inds(obj, year = NA, unit = NA, season = NA, area = NA,
  iter = NA)
}
\arguments{
\item{obj}{bla}

\item{year}{bla}

\item{unit}{bla}

\item{season}{bla}

\item{area}{bla}

\item{iter}{bla}
}
\value{
FLIBM object
}
\description{
Record state of population
}
