% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inds.R
\name{remove.inds}
\alias{remove.inds}
\title{Remove dead individuals}
\usage{
remove.inds(obj, year = NA, season = NA)
}
\arguments{
\item{obj}{bla}

\item{year}{bla}

\item{season}{bla}
}
\value{
FLIBM object
}
\description{
Remove dead individuals
}
