% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup.FLIBM.R
\name{spinup.FLIBM}
\alias{spinup.FLIBM}
\title{Spin up of inds in FLIBM object}
\usage{
spinup.FLIBM(obj, nyearsmax = 20, nyearsslope = 5, monitor = TRUE)
}
\arguments{
\item{obj}{FLIBM object}

\item{nyearsmax}{numeric. Maximum number of years for spin-up (Default=20)}

\item{nyearsslope}{numeric. Number of years for used to determin if spawning-
stock biomass is in equilibrium (via linear model of log SSB) (Default=5).}

\item{monitor}{logical. Should progression be printed.}
}
\value{
FLIBM object
}
\description{
Spin up of inds in FLIBM object
}
\examples{
stk <- create.FLIBM()
round(sum(stk$inds$wt * stk$inds$mat),3)
nrow(stk$inds)

stk <- spinup.FLIBM(stk)
round(sum(stk$inds$wt * stk$inds$mat, na.rm=TRUE),3)
nrow(stk$inds)



}
