% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLIBM.R
\name{trim.FLIBM}
\alias{trim.FLIBM}
\title{trim FLIBM dimensions}
\usage{
trim.FLIBM(
  obj,
  year = dimnames(obj$stock.a)$year,
  age = dimnames(obj$stock.a)$age,
  length = dimnames(obj$stock.l)$age,
  unit = dimnames(obj$stock.a)$unit,
  season = dimnames(obj$stock.a)$season,
  area = dimnames(obj$stock.a)$area,
  iter = dimnames(obj$stock.a)$iter
)
}
\arguments{
\item{obj}{FLIBM object}

\item{year}{bla}

\item{age}{bla}

\item{length}{bla}

\item{unit}{bla}

\item{season}{bla}

\item{area}{bla}

\item{iter}{bla}
}
\value{
FLIBM object
}
\description{
trim FLIBM dimensions
}
\examples{
data(stk1)
dimnames(stk1$stock.a)$year
stktmp <- trim.FLIBM(stk1, year = ac(1980:1985))
dimnames(stktmp$stock.a)$year
dimnames(stktmp$length.a)
dimnames(stktmp$age.l)


}
