% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLIBM.R
\name{window.FLIBM}
\alias{window.FLIBM}
\title{expand FLIBM object years}
\usage{
window.FLIBM(
  obj,
  start = dims(obj$stock.a)$minyear,
  end = dims(obj$stock.a)$maxyear
)
}
\arguments{
\item{obj}{FLIBM object}

\item{nyears}{numeric. Number of years to expand}

\item{numeric}{Ending year for window}
}
\value{
FLIBM object
}
\description{
expand FLIBM object years
}
\examples{
data(stk1)
dimnames(stk1$stock.a)$year
stktmp <- window.FLIBM(stk1, start = 1995, end = 1998)
plot(stktmp$stock.a@stock.n)


}
